Require texStorage or sized texImage to use RGBX format in GL.
The angle extensions that adds the RGBX format type only adds the sized
format in Angle. Thus we can't use texImage2D with an unsized format to
create a texture. However, if supported we can use texStorage or texImage
with a sized internal format. Thus we now require one of those to be
avaiable along with the original extension to claim support for RGBX

Bug: b/236571669
Change-Id: Ifb2ca1a66aa254e3b9d7e702e79a0cfe00773c05
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/566878
Commit-Queue: Greg Daniel <egdaniel@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>