[graphite] Fix SamplerState creation.
I was creating a descriptor wrapped in a smart pointer, then creating a
raw descriptor and setting up that, and then using the original
wrapped descriptor to create the sampler. As a extra bonus, because of
the raw descriptor this also introduced a memory leak.

I also added some useful labels for GPU capture.

Bug: skia:13118
Change-Id: I8a589b2ae82e05f4fcad5c31177b04d5dde0cd97
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/566882
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>