Always enable COLRv1 variations in testing
As seen in [1] setting this on a per-test basis does not seem to work
reliably when tests are run in batches. In DM testing we'd always prefer
to see variable COLRv1 results. And the flag is temporary and going to
be removed when variable CORLv1 ships.

[1] https://skia-review.googlesource.com/c/skia/+/563336

Bug: skia:13563
Change-Id: Id8628fd889df36597785692ded51f7f83c397536
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/567236
Reviewed-by: Ben Wagner <bungeman@google.com>
Auto-Submit: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Ben Wagner <bungeman@google.com>