[Graphite] Have each Recorder have its own SkSLCompiler.
SkSLCompiler is not threadsafe so each Recorder will need its own.
This will mean each compiler will have to handle the slow initial
startup on their first use. However, we hope to mitigate this if we can
share the modules that are created at compiler setup which is a large
part of the overhead.

Bug: b/239822490
Change-Id: Id58af02db378cbbae5ca1b7377fb60e384950f09
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/565500
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>