[bazel] Set up staged rollouts into G3 using defines.
This will let us do staged rollouts of API removal or improvements.
Primarily we will want to set defines, but if occasionally
we need to set copts, we can use the BASE_COPTS from
copts.bzl as well.

Overall, this cleans up the public.bzl - I hope to remove
the skia_select as it makes things hard to reason about for me.

base_linkopts and base_copts are no longer used.

Change-Id: I995cd47b09fbdcf02ccba8c961c15af2082df24f
Bug: skia:12541
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/568238
Reviewed-by: Ben Wagner <bungeman@google.com>