Reland "Replace StrikeRefs with SkStrikePromises"
This is a reland of commit bf6537f851fb0f9f24bc57f6e0754a57494e2aea

This happens in FM, and not DM because in DM, all the tests
share the global strike cache. And that allows the strike
for SkFont() to be placed in the cache. When FM runs this,
the strike cache is empty. And the descriptor lookup on the promise
fails. Make the strike, and keep a ref for the length of the test
to keep it alive.

Original change's description:
> Replace StrikeRefs with SkStrikePromises
>
> Change-Id: I68ee17e60776c3c3ca3dcbd98d734274625ed455
> Reviewed-on: https://skia-review.googlesource.com/c/skia/+/569197
> Reviewed-by: Robert Phillips <robertphillips@google.com>
> Commit-Queue: Herb Derby <herb@google.com>

Change-Id: Ie87ba76c1024e931cd1d75643087fc4ef57d8414
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/569740
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Herb Derby <herb@google.com>