[graphite] Transform bounding box on the GPU
Switches to instanced rendering and uploads either the transform or
the inverse transform and bounding box. The vertex shader then
calculates the other set of coordinates from the bounding box, so
regardless of fill-type, both device and local coords are available
in the shader to be (eventually) sent to the paint SkSL as a varying.

Bug: b/238761548
Change-Id: Idcb6b94ff457bc89b3c8410256a82d5d3fefa31a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/565008
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>