Remove UnresolvedFunction node.
Previously, lists of function overloads were stored in a dedicated
Symbol node called `UnresolvedFunction`. Whenever a new overload was
created, a new `UnresolvedFunction` was created to store the list of
candidates, and we had a large amount of bookkeeping code involved in
maintaining this candidate list.

Now, the list of candidates is stored as a linked list, and a
FunctionDeclaration simply points to the next overload candidate.
Places in the code which need to consider each overload candidate can
walk the chain directly. In practice, this removes a lot of complexity
from overload resolution, since we no longer need to treat overloads
as a special case; they just have a longer linked list.

There are some changes to const handling here which make this work
possible. In particular, transferring ownership of an object to a
SymbolTable no longer immediately makes it const. Constness is now
added when the object becomes a named/lookup-able symbol.

Change-Id: I3f4dc926bc9cb612ae86e65472b92119aa152d4e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/564460
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>