Reland "Use the remote canvas to convert Slugs"
This is a reland of commit 46bbcd3954d1f9a0c8b86cd27fc514fc5ee53690

In the dtor of the PathOpSubmitter, the code tried to use the
strike being null as a guard to check if the Paths had been
created. This does not work for the remote glyph cache, because
the promise can not produce the strike. Change to an explicit
guard.

Added a test that checks this condition.

Original change's description:
> Use the remote canvas to convert Slugs
>
> Add flag SK_SUPPORT_LEGACY_SLUG_CONVERT to move back to the old
> behavior if needed.
>
> The SkStrikePromise does everything that StrikeRef did, but in a more
> focused and testable manner. Take care nulling out the ref in
> the promise for the different SubRun types.
>
> Bug: chromium:1320305
> Bug: chromium:1320930
> Bug: chromium:1320153
> Change-Id: I9e30da4c17709f3b82b6ba89446af0a1502548f9
> Reviewed-on: https://skia-review.googlesource.com/c/skia/+/569323
> Commit-Queue: Herb Derby <herb@google.com>
> Reviewed-by: Robert Phillips <robertphillips@google.com>

Bug: chromium:1320305
Bug: chromium:1320930
Bug: chromium:1320153
Change-Id: I32a85aff361278b3736df1d8e30e835f5ef98b71
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/569747
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Herb Derby <herb@google.com>