Raster Pipeline: Use clang::musttail when possible
This forces tail call optimization on all stage functions. We still need
a stack management fallback strategy for other compilers, but this gives
us a way to simplify/optimize pipelines for our most common environment.

Note that emscripten will claim to support the attribute, but WASM
itself doesn't. As a result, the back-end will emit an error if we try
to use the attribute.

Change-Id: I86bed9f95b149fa0ed869e40edf49654424df100
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/569536
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>