[graphite] Pad the ends of uniform data blocks
This is necessary when we write uniform data to a shared storage buffer
that we index into with an attribute; the indexing assumes that buffer
entries will be aligned this way.

This is somewhat related to http://review.skia.org/556596, but since
this new fReqAlignment field is not used in SkSL generation (i.e.
ContextUtils.cpp's get_uniforms), I added it to the UniformManager class
rather than UniformOffsetCalculator.

Bug: b/242076321
Change-Id: Ifd591718a1556fb2d2afb1528a92cd06211fed83
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/569977
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>