Remove Inliner member variable from Compiler.
The Inliner is almost entirely stateless; the only state it needs is
"the context" and "how many things have been inlined" (strictly to keep
ExponentialGrowth.sksl in line). Since it was the original home of the
Mangler, it also held the responsibility of clearing the Mangler's
state, but that was more of an afterthought than a design goal.

I realized it's easier just to drop a fresh Inliner on the stack as
needed, rather than have a long-lived Inliner inside the Compiler that
we need to explicitly reset whenever it is used.

Change-Id: Ib003c61747a391c12a39551f830cecc8034d83e1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/571122
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>