On certain Adrenos running WebGL, flush before TexSubImage2D.
Something about WebGL seems to be tickling a driver bug on certain
Adrenos, such that the uploads for glyphs for the text renderer are
out of sync with the renders using those uploads. Flushing after
the upload doesn't fix it, but flushing before does.

Bug: b/229626353
Change-Id: I4a8e73e3189f60a1468467b6c861616e4ec00af8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/571418
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>