[graphite] Transform filled path control points on GPU
The tessellated curve and wedge variants for filled paths now
approximate the scale factor on the CPU to decide resolve level, but
the full transform is sent to the GPU as a uniform and we transform
to device space after tessellating.

This makes it possible to have both the local and device coords
available in the vertex shader if the paint needs local coords, and
it also is an important step in supporting perspective path rendering.
Later on we can adjust the scale approximation to be more local to
each path verb and incorporate the effects of perspective division.
But the uploaded control points still can be 2D, and we let the GPU
handle the final perspective division and w clipping.

Bug: b/238770428
Bug: b/238761548
Change-Id: I8ba017f9f98352f8162075b751589a2dfe64648b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/565338
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>