[bazel] Use newer config setting for string_flag (multiple)
The change from config.string [1] to config.string_list [2]
is that the former only supports setting one item as a default,
not multiple.

If, for example, we had wanted to set the default value of
//src/codec/include_decoder to be
['jpeg_decode_codec', 'png_decode_codec']
we could not without switching to string_list.

With Bazel 5.2.0, you could set the flags for string_list
exactly once with a flag (as a comma separated list).
In 5.3.0 they added the repeatable option which allows
us to set it multiple times and have it be aggregated.
As such, we move forward with the 5.3.0rc1 (which also
lets us provide feedback before release if necessary).

This also makes our flag macros align more with Bazel
style (flag_name -> name) and relocates them to a less
hidden spot.

See also http://cl/468514170

[1] https://bazel.build/rules/lib/config#string
[2] https://bazel.build/rules/lib/config#string_list

Change-Id: Ic0ef0de51078acc7212c43e9a58b0a121240b38e
Bug: skia:12541, b/237076898
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/571419
Reviewed-by: Ben Wagner <bungeman@google.com>