[graphite] Generate SkSL with SSBO lookups
This defines an SSBO index attribute name field on the RenderStep class.
If the field is null, we generate SkSL as we've been doing, but if it's
not null, we generate SkSL that indexes into storage buffers using the
attribute and an associated varying.

This is currently a no-op since the field is simply set to null.
Eventually we'd want to set it based on GPU caps. That might be a little
annoying since, as implemented now, we'd have to pass in the caps to
every RenderStep constructor. Maybe we could store the notion of
"using SSBOs" somewhere else if it seems more convenient.

In order to pass the SSBO index varying name all the way through to
getMangledUniformName, this stores the index name on the SkShaderInfo
object that's passed around various functions in SkShaderCodeDictionary.
Alternatively we could pass the index name around separately.

Bug: b/242076321
Change-Id: I4fb5a3c6353d5bd8903913564039a054e0fbb03e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/571096
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>