Optimize bicubic raster pipeline stages by sharing weights
We were re-computing each X and Y weight four times. That amounted to 32
evaluations of the bicubic_wts function. The new code does the
evaluation once (a total of 8 calls to bicubic_wts). The tradeoff is
having to store and load all the weight values to memory.

Change-Id: I8cffc3650ab981fea57c37d95322495f0bc86370
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/571637
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>