[bazel] Add rules to support jxl decoding
This was originally added in
https://skia-review.googlesource.com/c/skia/+/482596

I took the list of C++ source files in //third_party/libjxl/BUILD.gn
and supplemented them with the necessary header files.

The jxl codec depends on highway and brotli, but both of these
already have Bazel support which we can use out of the box.

A follow-on CL will add this codec to our "kitchen sink IWYU"
build and clean up any necessary headers.

See also http://cl/469238305

Change-Id: I1cfe4754c5b5f8208ba8a2435bbb19021e82e52c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/571935
Reviewed-by: Ben Wagner <bungeman@google.com>
Reviewed-by: Leon Scroggins <scroggo@google.com>