[PathOps] Don't offset curves when sorting angles
At intersections, when sorting into angular sections, the old code
would translate all control points in a curve if it's first point
didn't match the intersection exactly. This changes it to only move
the first control point of the curve.

The motivation for this is intuitive, it seems to preserve the overall
convex hull of the curve better in relation to the others, and it
fixes two cases of Simplify() failing glyphs from the Noto font.

This CL also updates the path ops debug code to compile when that's
enabled since it had bit-rotted. Adds unit tests for the failing Noto
glyphs. The cases from comment 7 (_b) and comment 15 (_c) now pass
with the change to SkOpAngle. The case from comment 5 (_a) still does
not pass but that will be explored in another CL.

This change may be addressing some of the same issues that the old,
but reverted and never-relanded, https://skia-review.googlesource.com/c/skia/+/7373/ as trying to handle.

Bug: skia:11958
Change-Id: Iac6f84e93bbc7f5fe702f3bcfa3c1824558d6819
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/565997
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>