Fix large emoji scaling
At the point the code generates the StrikeSpec for the SubRun,
the gaugeMatrix is the original CTM, now the strike will
have a factor of the CTM and the canvas will have a factor of the
CTM. This causes the CTM to be applied twice. The StrikeSpec
should be created with the identity matrix.

Change-Id: I5a180831a97db06502995f128e1e3d1d2672600e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/572286
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Herb Derby <herb@google.com>