Expand layer to clip for color filters/blenders that affect transparent black
If a saveLayer had a restore paint with a color filter that affected
transparent black, or if it's blend mode modified transparent black,
the final rendering would visually leak any restriction on the layer
size that we calculated based on the user bounds hint.

We work to avoid this leakage when there are backdrop filters and if
there is an image filter that affects transparent black, but the CF
and blender of the restore paint weren't being considered.

Chromium needs some slight rebaselines. Flutter's display list logic
needs to be fixed, and their layer widgets need to be updated to stop
relying on the user bounds as an implicit clip.

Bug: skia:12083, skia:12303
Change-Id: Idc49ba146635d0c0879be61c4ca676e16460b849
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/570779
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>