Remove obsolete caps bit for node pools.
We also have a ProgramSettings flag for disabling node pools, so the
caps bit was largely redundant. It looks like it was used to run an
experiment in February 2021 (unexplained memory corruption on Adreno
530) but it seems that the corruption issue disappeared on its own
before the experiment rolled out anyway, when Chrome 88 landed.

This is mostly helpful because we want SkSL to avoid relying on caps
bits outside of `convertProgram`.

Change-Id: I4878a25908dee0bf088d0412564dbe346563f014
Bug: chromium:1147008, chromium:1164271
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/572292
Commit-Queue: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>