Preserve Settings expressions during Rehydration.
If we convert Setting expressions to a Literals at rehydration time,
the module becomes specialized to one specific ShaderCaps
configuration. Instead, we now keep the Setting expression as-is; that
way, the module can be reused by every compiler.

This change affects all of our backends, because Programs can contain
functions from the shared-element table, and a shared function might
now contain a Setting. We could dodge this by cloning the function
and rewriting it, but that's a _very_ heavy hammer, so for now we
don't do this. Instead, codegens now allow Setting expressions to
exist, and convert them to literals when they are found, using the
shader caps from the Program's context.

We can still replace a Setting expression with the equivalent literal
during the inlining phase. In practice, this does a decent job of
optimizing away many Setting expressions; it still generates slightly
different code than before, but hopefully in a way that will look the
same to a driver.

Change-Id: I2f8618ce79207b990d02220ecea75ddf601b5dde
Bug: skia:13671
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/571700
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>