[COLRv1] Handle zero color stop range in interpolated gradients
When the color stops are in the same position, for extend modes repeat
and reflect we cannot render anything meaningful. For extend mode pad,
no interpolation should be performed, as the scale factor cannot be
computed. But instead, use the default behavior of the shader, which
does the right thing in this situation.

Bug: skia:50557
Cq-Include-Trybots: luci.skia.skia.primary:Test-Android-Clang-GalaxyS20-GPU-MaliG77-arm64-Release-All-Android_NativeFonts
Change-Id: I55bd3a4c9e0878bcc462695b413b6133a149ed26
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/572934
Commit-Queue: Dominik Röttsches <drott@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>