Remove state from BuiltinMap class.
Previously, BuiltinMap tracked the inclusion of elements into the
program directly, and this state would be cleared every time an SkSL
ThreadContext was initialized. This became a thread hazard when
modules could be reused across multiple Compilers.

The code which inserted elements from the BuiltinMap was already
tracking the array of inserted-elements on its own, so we no longer
actually needed the BuiltinMap to track this state anyway. (This
wasn't the case when the BuiltinMap was originally written.)

Change-Id: I338cafcf5fefef2363ca19512bd8f3757532eb28
Bug: skia:13666
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/573015
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>