Check if sampler index is greater than num of samplers in GL.
We are crashing with a nullptr in SamplerObjectCache when trying to
access an fTextureUnitStates at a specific index. The guess is that we
may be getting into a state where we try to bind a sampler after we've
abandon things. So this check hopefully catches that avoids the crash.

This fix does not fix the underlying issue of why we are getting down
into here if we did abandon things. Most likely we are not abandoned
when we go through the higher level checks but something triggers the
abandon in the middle of the call stack after we think things have
been check and safe. Just given the call stacks of the crash this is
not obvious where that would be. Thus the simple stop gap for now.

Bug: chromium:1356293
Change-Id: Idee973cece3e86b8ad0e7969821100ec113b2123
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/573642
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>