Use matrices instead of font size for controlling large glyphs
When a glyph exceeds the size of the atlas, we make the glyph
smaller by adjusting the font size. This change keeps the font size
the same, which has additional benefits, and creates the glyphs based
on the CTM which has been scaled down.

A new TransformedMaskVertexFiller2 takes the creation matrix and
uses that to produce the file quads. In order to map the glyph data
it creates a viewDifference matrix:

   viewDifference = [CTM] * [creationMatrix]^-1.

In addition, the handling of sizing the glyphs in perspective
is updated to use Jim's method for calculating a good scale size
for the glyphs.

Bug: skia:12121
Bug: chromium:1303466
Bug: chromium:1116365
Change-Id: I200e0285720c5fa53fc02e7015f5946f7246eadf
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/573560
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>