Complete the implementation of AsWinding
(an update of https://skia-review.googlesource.com/c/skia/+/497140)

(Matthew Blanchard's original comment)

Changed aswinding to use the signed area of the polygon defined by the
contour's control points instead of the derivative of the left edge to
determine winding direction.

(Cary Clark's comment)
AsWinding() was never fully implemented at the time that PathOps was
destaffed. It included a debug assert that was triggered when the
incomplete code was encountered. The old code look at the slope of the
leftmost edge of each contour to determine if the contour as a whole
is clockwise or anticlockwise. If the slope of that edge could not
be determined (for instance, if it was horizontal, or a point), the
code asserted and failed.

This CL by Matthew Blanchard remedies this by considering all of the
edges of each contour. Treating the entire contour as lines between
control points, it measures whether each triangle formed contributes
and positive or negative value as it turns clockwise or anticlockwise.
The sum of those values gives the overall contour direction.

This code is a simplification of actual area calculation, as it does
not consider the curve type. However, for all examples given in the
bug postings, it correctly reorders the contours so that the xor fill
version of the path draws the same as the winding fill version of
the shape. This is verified by using pathops to xor the two together,
which on success returns an empty path.

As Matthew cannot be located, I took his CL and simplified the
original slightly to reduce size of change.

I added tests extracted from bugs 12040 and 13496.

Bug: skia:12040, skia:13496, skia:13498
Change-Id: Ia45afb6ffd662ca1b367ed5ffc2ccfdfdf9f50c9
R=herb@google.com
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/572958
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: Herb Derby <herb@google.com>