Rename DSLParser to Parser.
This is the only parser we have. We already have lots of classes with
DSL in the name, or in namespace "dsl"; having the prefix on every
Parser method adds visual clutter without giving signal to the reader.

Change-Id: I6ab86227ba0198e6e7fd65b666c91f458e136fee
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/573647
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>