[graphite] Memoryless MSAA color/depth/stencil attachments
Switches the MSAA color attachment and the depth/stencil attachments
to memoryless. Adds logic in MtlCommandBuffer to draw the resolve
texture to the MSAA attachment as a "load" op. The MtlResourceProvider
caches MtlGraphicsPipelines for handling the draw call, defined as
static MSL internally.

In order for MtlCommandBuffer to create these pipelines, the
QueueManager provides the ResourceProvider to it when the CB is made.
Vulkan will need access to the resource provider as well.

Change-Id: I7d149d6342e66513242f74441ea0cd7b4d762e24
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/574196
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>