Experiment: remove Inliner from SkSL.
This saves ~25K of binary size. There are known downsides (see
http://go/optimization-in-sksl-inliner) but it's possible that
the benefits outweigh the downsides.

Current plan: land this briefly, check out the impact for 1-2 days,
and then assess the situation. For now let's assume we will roll it
back, but there's always the off chance it's perfect as-is.

Change-Id: Iefc8fd2789e9a2f988486cd2ba1a4488af9c5a7c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/569744
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>