Disable polyfills by default in ShaderCaps.
Typically, our caps bits default to the straightforward "no
workarounds" case. The polyfill caps were an exception; in the default
config, the GLSL code generator would rewrite `fma` and `determinant`
into their arithmetic forms. This was unintuitive and it meant we
needed to manually toggle these flags when we wanted the results to be
as close to the original as possible.

This CL inverts the default state of the polyfill caps bits. This
affects a few of our test files (for the better!) because they are
built in the Standalone configuation (i.e., all flags default).
GrGLCaps::initGLSL will always initialize the flag values instead
of relying on their defaults.

Interestingly, `cross` in sksl_gpu changed its behavior based on  sk_Caps.builtinDeterminantSupport. The Caps bit was previously only
set in GLSL and ignored in VkCaps and MtlVaps. This meant that until
now, `cross` wouldn't use the built-in determinant intrinsic in
SPIR-V or Metal. (!!)

Change-Id: I0ec560e686bd636e45bc33f930fe8ca24c5d518f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/572665
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>