[graphite] Add unique ID to RenderStep, simplify GraphicsPipelineDesc
Adds a 32-bit unique ID to RenderStep so that GraphicsPipelineDesc can
just be (effectively) two uint32_t's. Removes redundant key storage,
and simplify interface.

Simplifies MtlCaps and other impacted areas.

The one place that has increased complexity is when actually building
the SkSL for a pipeline, which requires recovering the real RenderStep
from the ID on the GraphicsPipelineDesc. This is fine since it's a very
rare occurrence and the RendererProvider is already available off of the
SharedContext.

Change-Id: I62db44bc6f95f22b66e3b56865ac1cae201b67e6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/574337
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>