Remove the origin offset from the SkGlyphRunList bounds
Having the sourceBounds() offset by the SkGlyphRunList's origin
made it difficult to get the unshifted bounds, and hard to remember
which bounds to use with which matrix.

I have changed it so that sourceBounds() is unoffset, and behaves like
SkTextBlob's bounds(). I have added sourceBoundsWithOrigin() to
include the bounds + the origin offset. I went through the code
checking all sourceBounds uses, and updated or fixed them.

Here are the calls to use with the different matrices:
* drawingMatrix use sourceBoundsWithOrigin
* positionMatrix use sourceBounds

Change-Id: I9e87816431f8b6fe95838f46d5e54e555be2ec17
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/575499
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Herb Derby <herb@google.com>