Manually specify Bazel exporter rule mapping.
This new implementation relies on the caller to explicitly
specify the *.gni files to create, the Bazel rules to
export, and to which GN file list the rule files should be
written.

This greatly simplifies the implementation and use as the
prior impl relied on a more complicated system of mapping
rule folders to *.gni files, which was insufficient and
needed a complicated set of merge/rename rules to
compensate for that approximate rule mapping.

Bug: skia:13485
Change-Id: I894ecd12806f215faac3f273a7a830d6c4dfc6b1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/575518
Reviewed-by: Kevin Lubick <kjlubick@google.com>