Bazel: Switch to `bazel query` for the GNI export.
Switch to `bazel query` as it is unaware of the build platforms as
well as the Skia build flags (`--enable_pdf_backend`, etc.).
This makes it a better candidate for the GNI exporter which needs
to create a set of *.gni files that work on all platforms.

The effect of this switch is that `select(...)` calls behave
as though they are always true - selecting the conditional
properties. If individual files are being selected then they
will be exported to the *.gni files - which may be desired.
If not desired then the solution is to move those files to
a new filegroup and instead select the filegroup. Selected
filegroups will be incorporated into the referencing rule
during a build, but will be ignored during the GNI export.

Bug: skia:13485
Change-Id: I046d2d6eb14ba01ba59ff6dcb313663245e4c5a7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/575976
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>