Adjust clamping behavior in raster pipeline
This brings RP behavior in line with SkVM and GPUs,
being less opinionated about colors being "in-gamut"
before the blend step.

This produces one bad GM result, in the readpixels
GM. That image is exhibiting a problem that exists
today -- if you feed a premul 8888 bitmap with colors
that are out-of-gamut to the legacy sprite blitters,
they will break and produce garbage. This is shown
directly in this CL:
  https://skia-review.googlesource.com/c/skia/+/580076

I have a start of a fix for that here:
  https://skia-review.googlesource.com/c/skia/+/579280

... but the fix doesn't cover the non-SIMD sprite
blitter. Regardless, I think this is just exposing an
issue that already exists, and one that is VERY rare
in normal real-world usage. On the flip-side, landing
this CL fixes incorrect results when using RP to draw
gradients interpolated in fancy new CSS color spaces.

Bug: skia:13715
Bug: skia:13108
Change-Id: I697050b7cffb2a7af1507da1e82898fa255f0d26
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/577901
Reviewed-by: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>