Add tests for SkBitmapProcState_matrixProps::pack
In an effort to fix the linked bug, this adds some tests for
three uses of pack (clamp, mirror, and repeat), which
has previously only been tested indirectly by certain
shaders on the CPU-backend (with SK_ENABLE_LEGACY_SHADERCONTEXT
defined).

To expose this private static function for testing, this CL
adds an sktests namespace and exposes the functions through
that. The exposed functions are real combinations used
(see [1]).

This CL also adds tests for the unpacking of the packed value.

[1] https://github.com/google/skia/blob/1f193df9b393d50da39570dab77a0bb5d28ec8ef/src/core/SkBitmapProcState_matrixProcs.cpp#L250-L261

Change-Id: I74001857d3469b1c16eea9115867a6ca79c383d1
Bug: 1357122
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/573880
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Herb Derby <herb@google.com>