[graphite] Simplify texture binding writing
Moves SamplerDesc to DrawTypes.h and has SkTextureDataBlock use that
instead of being a clone of its fields + a TextureProxy.

Adds a bindDeferredTexturesAndSamplers() function to
DrawCommands::List that creates space for the indices and returns that
to the caller to fill in, which removes the awkward maxNumTextures
tracking and intermediate copy that used to happen.

Change-Id: I3c0900aa240e7f83cd2447ca50565b2a0a7f58cf
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/573883
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>